+++ single.php
@@ -1,7 +1,14 @@
 <?php get_header(); ?>
+    <?php $use_zenwriter_formats = get_theme_mod('zenwriter_setting_zenwriter-post-format'); ?>
     <?php if (have_posts()) : while (have_posts()) : the_post(); ?>
         <article id="post-<?php the_ID(); ?>" <?php post_class('post post--full'); ?>>
-            <?php get_template_part('/template-parts/post-format-' . (get_post_format() ? : 'standard')); ?>
+            <?php
+            if ($use_zenwriter_formats) {
+                get_template_part('/template-parts/post-format-' . (get_post_format() ?: 'standard')); 
+            } else {
+                get_template_part('/template-parts/post-format-standard'); 
+            }
+            ?>
         </article><!-- .post -->
         <?php get_template_part('/template-parts/post-navigation'); ?>
         <?php comments_template(); ?>